requirements.txt
discord.py
python-dotenv
jishaku==2.5.1
requests~=2.31.0
pytz~=2022.2.1
aiohttp~=3.9.3
asyncpg